requirements.txt
python-telegram-bot==21.6
httpx==0.27.0
python-dotenv==1.0.1
fastapi==0.112.1
uvicorn[standard]==0.30.5